Fixing deleting account bug

+++ app/controllers/accounts_controller.rb
@@ -56,9 +56,11 @@ class AccountsController < ApplicationController
     @account = Account.find(params[:id])
 
     respond_to do |format|
-      transaction = Amount.where(account_id: Account.last.id)[0].transaction
-      Amount.where(transaction_id: transaction.id).destroy_all
-      transaction.destroy
+      transaction = Amount.where(account_id: @account.id)[0].try(:transaction)
+      if(transaction)
+        Amount.where(transaction_id: transaction.id).destroy_all
+        transaction.destroy
+      end
       if(@account.destroy)
         flash[:notice] = 'Account was successfully destroyed.'
